validate visualisations able to be added to dashboard #1826
